+++ MudBlazor.Extensions/Components/MudExMessageDialog.razor.cs
@@ -1,6 +1,9 @@
 ﻿using Microsoft.AspNetCore.Components;
 using MudBlazor.Extensions.Attribute;
+using MudBlazor.Extensions.Core;
 using MudBlazor.Extensions.Helper;
+using System;
+using System.Reflection;
 
 namespace MudBlazor.Extensions.Components;
 
@@ -9,7 +12,14 @@ namespace MudBlazor.Extensions.Components;
 /// </summary>
 public partial class MudExMessageDialog
 {
+    private ComponentBase _component;
     [CascadingParameter] MudDialogInstance MudDialog { get; set; }
+
+    /// <summary>
+    /// The color of the dialog icon for when the dialog is used in relation to a particular color theme.
+    /// </summary>
+    [Parameter]
+    public MudExColor IconColor { get; set; } = Color.Error;
 
     /// <summary>
     /// Gets or sets the class for the content of the dialog
@@ -89,11 +99,79 @@ public partial class MudExMessageDialog
     [Parameter, SafeCategory("Data")]
     public double ProgressMax { get; set; } = 100.0;
 
+    /// <summary>
+    /// If true StateHasChanged is not happen
+    /// </summary>
+    [Parameter, SafeCategory("Behavior")]
+    public bool PreventStateHasChanged { get; set; }
+
+    /// <inheritdoc />
+    protected override void OnAfterRender(bool firstRender)
+    {
+        if(firstRender && Buttons?.Any(b => b.HasCondition) == true)
+            UpdateConditions();
+        base.OnAfterRender(firstRender);
+    }
+
+    private void UpdateConditions()
+    {
+        foreach (var btn in Buttons.Where(b => b.RenderedReference != null && b.HasCondition))
+            btn.RenderedReference.SetDisabled(!btn.CanExecuted(Component));
+    }
+
+    /// <inheritdoc />
+    protected override bool ShouldRender() => !PreventStateHasChanged && base.ShouldRender();
 
     /// <summary>
     /// Gets or sets the component associated with the dialog
     /// </summary>
-    internal ComponentBase Component { get; set; }
+    internal ComponentBase Component
+    {
+        get => _component;
+        set
+        {
+            if(_component == value) return;
+            _component = value;
+            if (_component != null && Buttons.Any(b => b.HasCondition))
+                AdjustEventCallbacks(_component); // Attach Event callbacks to update conditions for buttons
+        }
+    }
+
+    private void AdjustEventCallbacks(ComponentBase component)
+    {
+        if(component == null) return;
+        var eventCallbackProperties = component.GetType().GetProperties()
+            .Where(p => p.PropertyType.IsGenericType &&
+                        p.PropertyType.GetGenericTypeDefinition() == typeof(EventCallback<>));
+
+        foreach (var prop in eventCallbackProperties)
+        {
+            var existingCallback = prop.GetValue(component);
+            if (existingCallback == null) continue;
+
+            var callbackType = prop.PropertyType.GetGenericArguments().FirstOrDefault();
+            if (callbackType == null) continue;
+
+            var method = GetType().GetMethod(nameof(WrapCallback), BindingFlags.NonPublic | BindingFlags.Instance)?.MakeGenericMethod(callbackType);
+            if(method == null) continue;
+            var newCallback = method.Invoke(this, new[] { existingCallback });
+            prop.SetValue(component, newCallback);
+        }
+    }
+
+    private EventCallback<T> WrapCallback<T>(EventCallback<T> existing)
+    {
+        var wrappedCallback = EventCallback.Factory.Create<T>(
+            Component,
+            (value) =>
+            {
+                existing.InvokeAsync(value);
+                UpdateConditions();
+            });
+
+        return wrappedCallback;
+    }
+
 
 
     void Submit(DialogResult result)