+++ MudBlazor.Extensions/Options/HelperTextAlignment.cs
@@ -1,0 +1,8 @@
+﻿namespace MudBlazor.Extensions.Options;
+
+public enum HelperTextAlignment
+{
+    Default,
+    OnComponent,
+    OnLabel,
+}